+++ packages/harpstrata/src/harpstrata/Degree/index.ts
@@ -16,3 +16,7 @@ export const getDegreeMatrix = (halfstepIndexMatrix: HalfstepIndexMatrix, root: 
     })
   })
 }
+
+export type { Degree, DegreeRow, DegreeMatrix } from './types'
+export { DegreeIds } from './types'
+export { ROOT, FLAT2, SECOND, FLAT3, THIRD, FOURTH, FLAT5, FIFTH, FLAT6, SIXTH, FLAT7, SEVENTH } from './constants'