+++ packages/harpstrata/src/harpstrata/IsActive/getIsActiveMatrix/types/index.ts
@@ -1,0 +1,7 @@
+export enum IsActiveIds {
+  Active = 'ACTIVE',
+  Inactive = 'INACTIVE',
+}
+
+export type IsActiveRow = ReadonlyArray<IsActiveIds | undefined>
+export type IsActiveMatrix = ReadonlyArray<IsActiveRow>