+++ packages/harpstrata/src/harpstrata/Pozition/index.ts
@@ -7,3 +7,4 @@ pozitionMap.set(SECOND.id, SECOND)
 
 export const getActivePozitionIds = (): PozitionIds[] => Array.from(pozitionMap.keys())
 export const getPozition = (pozitionId: PozitionIds): Pozition => pozitionMap.get(pozitionId)
+export type { PozitionIds } from './types'