+++ packages/harpstrata/src/harpstrata/index.ts
@@ -1,9 +1,10 @@
 import { getActivePozitionIds, getPozition } from './Pozition'
 import { getPitchMatrix, getActivePitchIds, getPitch } from './Pitch'
+import { getIsActiveComplex } from './IsActive'
 import { getDegreeMatrix } from './Degree'
 import { getActiveApparatusIds, getApparatus } from './Apparatus'
 
-import type { HarpStrata, PozitionIds, ApparatusIds, PitchIds } from './types'
+import type { HarpStrataProps, HarpStrata } from './types'
 
 
 export const getApparatusIds = getActiveApparatusIds
@@ -12,17 +13,26 @@ export const getPozitionIds = getActivePozitionIds
 
 export const getPitchIds = getActivePitchIds
 
-export const getHarpStrata = (apparatusId: ApparatusIds, pozitionId: PozitionIds, keyPitchId: PitchIds): HarpStrata => {
+export const getHarpStrata = (props: HarpStrataProps): HarpStrata => {
+  const { apparatusId, pozitionId, keyPitchId, activeIds } = props
+
   const apparatus = getApparatus(apparatusId)
   const pozition = getPozition(pozitionId)
   const pitch = getPitch(keyPitchId)
-  const degreeMatrix = getDegreeMatrix(apparatus.halfstepIndexMatrix, pozition.root)
-  const pitchMatrix = getPitchMatrix(apparatus.halfstepIndexMatrix, pitch.id)
+
+  const { halfstepIndexMatrix } = apparatus
+  const { root: pozitionRoot } = pozition
+  const { id: pitchId } = pitch
+
+  const degreeMatrix = getDegreeMatrix(halfstepIndexMatrix, pozitionRoot)
+  const pitchMatrix = getPitchMatrix(halfstepIndexMatrix, pitchId)
+  const isActiveComplex = getIsActiveComplex({degreeMatrix, pitchMatrix, activeIds: activeIds })
 
   return {
     apparatus,
     degreeMatrix,
     pitchMatrix,
+    isActiveComplex,
   }
 }
 
@@ -46,3 +56,6 @@ export {
   PitchIds,
   Pitch
 } from './types'
+
+export type { ActiveDegreeIds, ActivePitchIds, ActiveIds, IsActiveMatrix, IsActiveRow, IsActiveComplex } from './IsActive'
+export { IsActiveIds } from './IsActive'