+++ packages/harpstrata/src/harpstrata/types.ts
@@ -1,11 +1,21 @@
-import type { PitchMatrix } from './Pitch/types'
-import type { DegreeMatrix } from './Degree/types'
-import type { Apparatus } from './Apparatus/types'
+import type { PozitionIds } from './Pozition'
+import type { PitchIds, PitchMatrix } from './Pitch'
+import type { ActiveIds, IsActiveComplex } from './IsActive'
+import type { DegreeMatrix } from './Degree'
+import type { Apparatus, ApparatusIds } from './Apparatus'
+
+export type HarpStrataProps = {
+  readonly apparatusId: ApparatusIds;
+  readonly pozitionId: PozitionIds;
+  readonly keyPitchId: PitchIds;
+  readonly activeIds: ActiveIds;
+}
 
 export interface HarpStrata {
   readonly apparatus: Apparatus;
   readonly degreeMatrix: DegreeMatrix;
   readonly pitchMatrix: PitchMatrix;
+  readonly isActiveComplex: IsActiveComplex;
 }
 
 export {