+++ packages/harpstrata/CHANGELOG.md
@@ -12,7 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Security: to invite users to upgrade in case of vulnerabilities.
 
 
-## [Unreleased](https://github.com/js-jslog/harpstrata/compare/v3.0.2...HEAD) - yyyy-mm-dd
+## [Unreleased](https://github.com/js-jslog/harpstrata/compare/v4.0.0...HEAD) - yyyy-mm-dd
+## [v4.0.0](https://github.com/js-jslog/harpstrata/releases/tag/v4.0.0) - xxxx-xx-xx
+### Added
+- MAJOR: Add IsActiveComplex to the HarpStrata type
+### Changed
+- MAJOR: Change getHarpStrata function signature to typed object
+
 ## [v3.0.2](https://github.com/js-jslog/harpstrata/releases/tag/v3.0.2) - 2020-05-03
 - PATCH: Include Pitch and PitchIds types in the entrypoint file
 - PATCH: Fix version links in CHANGELOG for `v3.0.1` & `v3.0.0`